KSY-알림장_전체조회_로직_수정

+++ src/main/java/com/aico/aibayo/entity/ClassEntity.java
@@ -17,7 +17,7 @@ public class ClassEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "class_no")
-    private int classNo;
+    private Long classNo;
     @Column(name = "class_name")
     private String className;
     @Column(name = "class_age")
